+++ src/utils.js
@@ -1,0 +1,7 @@
+const getRandomInteger = (min, max) => {
+  return Math.floor(min + Math.random() * (max + 1 - min));
+};
+
+const getRandomValue = (array) => array[Math.floor((Math.random() * array.length))];
+
+export {getRandomInteger, getRandomValue};
